requirements/local.txt
asgiref==3.8.1
Django==5.1.1
django-decouple==2.1
django-filter==24.3
djangorestframework==3.15.2
Markdown==3.7
pillow==10.4.0
psycopg2-binary==2.9.9
python-decouple==3.8
sqlparse==0.5.1
Unipath==1.1
